FDB-419 fix test

+++ tests/regressions/FDB-419/fdb_419_regression_test.cc
@@ -21,6 +21,7 @@
 #include <csignal>
 #include <fstream>
 #include <sstream>
+#include <libproc.h>
 
 using eckit::Log;
 using eckit::PathName;
@@ -44,7 +45,7 @@ void spawn_reaper(const pid_t child_pid) {
         throw std::runtime_error(ss.str());
     }
     if (pid > 0) {
-        Log::info() << "Started reaper process" << pid << std::endl;
+        Log::info() << "Started reaper process " << pid << std::endl;
         return;
     }
     setsid();
@@ -129,14 +130,35 @@ pid_t run_server(const PathName& fdb_server_path, const PathName& log_file) {
     return {};
 }
 
-void kill_server(const pid_t pid) {
-    int rc = kill(pid, SIGKILL);
-    Log::info() << "Killing " << pid << std::endl;
-    if (rc == -1) {
-        Log::error() << "Could not kill fdb-server: " << std::strerror(errno) << std::endl;
+void kill_server(const pid_t ppid) {
+
+    int n = proc_listpids(PROC_ALL_PIDS, 0, NULL, 0);
+    int *buffer = (int *)malloc(sizeof(int)*n);
+
+	int numpids = proc_listpids(PROC_PPID_ONLY, (uint32_t)ppid, buffer, n*sizeof(int));
+	if (numpids != -1) {
+		numpids /= sizeof(int);
+        for (int i=0; i < numpids; i++) {
+            pid_t pid = buffer[i];
+            if (pid == 0 || pid == ppid) continue;
+            int rc = kill(pid, SIGKILL);
+            EXPECT_EQUAL(rc, 0);
+            Log::info() << "Killing child " << pid << std::endl;
+            if (rc == -1) {
+                Log::error() << "Could not kill fdb-server child: " << std::strerror(errno) << std::endl;
+            }
+        }
     }
-    std::this_thread::sleep_for(1s);
+    free(buffer);
+
+    int rc = kill(ppid, SIGKILL);
     EXPECT_EQUAL(rc, 0);
+    Log::info() << "Killing parent " << ppid << std::endl;
+    if (rc == -1) {
+        Log::error() << "Could not kill fdb-server parent: " << std::strerror(errno) << std::endl;
+    }
+
+    std::this_thread::sleep_for(1s);
 }
 
 size_t count_in_file(const PathName& file, const std::string& text) {
@@ -168,13 +190,13 @@ CASE("FDB-419") {
     const auto req = fdb5::FDBToolRequest::requestsFromString("class=rd,expver=xxxx")[0];
     EXPECT_NO_THROW(fdb_a.list(req));
     EXPECT_NO_THROW(fdb_b.list(req));
-    kill_server(-fdb_server_pid);
+    kill_server(fdb_server_pid);
     EXPECT_THROWS(fdb_a.list(req));
 
     fdb_server_pid = run_server(fdb_server_path, "srv2.log");
     EXPECT_NO_THROW(fdb_a.list(req));
     EXPECT_NO_THROW(fdb_b.list(req));
-    kill_server(-fdb_server_pid);
+    kill_server(fdb_server_pid);
     // Ensure only one connection is established with each server
     EXPECT_EQUAL(count_in_file("srv1.log", "FDB forked pid"), 1);
     EXPECT_EQUAL(count_in_file("srv2.log", "FDB forked pid"), 1);
